# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dark Realm RPG API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SaveGameRequest {
  playerId: string;
  playerName: string;
  /** JSON-serialized game state */
  gameState: string;
}

export interface SaveGameResponse {
  success: boolean;
  savedAt: string;
}

export interface LoadGameResponse {
  found: boolean;
  playerName?: string;
  gameState?: string;
  savedAt?: string;
}

export interface LeaderboardEntry {
  rank: number;
  playerName: string;
  level: number;
  dungeonFloor: number;
  totalKills: number;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
}

export type LoadGameParams = {
  playerId: string;
};