# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dark Realm RPG API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Save game state
 */
export const SaveGameBody = zod.object({
  playerId: zod.string(),
  playerName: zod.string(),
  gameState: zod.string().describe("JSON-serialized game state"),
});

export const SaveGameResponse = zod.object({
  success: zod.boolean(),
  savedAt: zod.string(),
});

/**
 * @summary Load game state
 */
export const LoadGameQueryParams = zod.object({
  playerId: zod.coerce.string(),
});

export const LoadGameResponse = zod.object({
  found: zod.boolean(),
  playerName: zod.string().optional(),
  gameState: zod.string().optional(),
  savedAt: zod.string().optional(),
});

/**
 * @summary Get top players leaderboard
 */
export const GetLeaderboardResponse = zod.object({
  entries: zod.array(
    zod.object({
      rank: zod.number(),
      playerName: zod.string(),
      level: zod.number(),
      dungeonFloor: zod.number(),
      totalKills: zod.number(),
    }),
  ),
});