# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dark Realm RPG API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  rank: number;
  playerName: string;
  level: number;
  dungeonFloor: number;
  totalKills: number;
}