# lib/api-zod/src/generated/types/saveGameRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dark Realm RPG API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveGameRequest {
  playerId: string;
  playerName: string;
  /** JSON-serialized game state */
  gameState: string;
}